+++ node/risk-app/server/services/__tests__/address.js
@@ -1,7 +1,7 @@
 const addressService = require('../address')
 
-describe('/Map page test', () => {
-  it('Assert Map page', async () => {
+describe('Address service', () => {
+  it('capitaliseAddress function capitalises the first of each letter other than postcode which remains in capitals', async () => {
     const address = '28, NORTHFIELD CLOSE, NEWPORT, NP18 3EZ'
     expect(addressService.capitaliseAddress(address)).toEqual('28, Northfield Close, Newport, NP18 3EZ')
   })
